requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
comtypes==1.1.9
COVID19Py==0.3.0
docopt==0.6.2
geojson==2.5.0
idna==2.10
Js2Py==0.70
MouseInfo==0.1.3
numpy==1.20.2
packaging==20.9
Pillow==8.1.2
pipwin==0.5.1
PyAudio @ file:///C:/Users/denux/pipwin/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
pyjokes==0.6.0
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyowm==3.2.0
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyPrind==2.11.2
PyRect==0.1.4
PyScreeze==0.1.26
pySmartDL==1.3.4
PySocks==1.7.1
python-dotenv==0.17.1
pyttsx3==2.90
PyTweening==1.0.3
pytz==2021.1
pywhatkit==3.9
pywin32==300
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
tzlocal==2.1
urllib3==1.26.4
wikipedia==1.4.0
